Created User Profile Dtos and started work on endpoint

+++ Backend/src/main/java/com/pz/designmatch/repository/CommissionRepository.java
@@ -3,5 +3,5 @@ package com.pz.designmatch.repository;
 import com.pz.designmatch.model.Commission;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CommissionRepository extends JpaRepository<Commission, Integer> {
+public interface CommissionRepository extends JpaRepository<Commission, Long> {
 }
