+++ Backend/src/main/java/com/pz/designmatch/repository/CompanyProfileRepository.java
@@ -3,5 +3,5 @@ package com.pz.designmatch.repository;
 import com.pz.designmatch.model.user.CompanyProfile;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CompanyProfileRepository extends JpaRepository<CompanyProfile, Integer> {
+public interface CompanyProfileRepository extends JpaRepository<CompanyProfile, Long> {
 }
